Decompose components.rb
This helps to prevent circular dependencies, and makes the dependencies
on components more clear.
