+++ lib/fauxpaas/components/backend_runner.rb
@@ -1,0 +1,10 @@
+require "fauxpaas/cap_runner"
+require "fauxpaas/components/system_runner"
+
+module Fauxpaas
+  class << self
+    def backend_runner
+      @backend_runner ||= CapRunner.new(system_runner)
+    end
+  end
+end
