+++ lib/fauxpaas/components/instance_repo.rb
@@ -1,0 +1,11 @@
+require "fauxpaas/file_instance_repo"
+
+module Fauxpaas
+  class << self
+    def instance_repo
+      @instance_repo ||= FileInstanceRepo.new(instance_root)
+    end
+
+    attr_writer :instance_repo
+  end
+end
