+++ lib/fauxpaas/components/token.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+require "fauxpaas/components/paths"
+
+# Fake Platform As A Service
+module Fauxpaas
+
+  class << self
+    def split_token
+      @split_token ||= File.read(root + ".split_token").chomp.freeze
+    end
+  end
+
+end
