+++ lib/fauxpaas/deploy_config.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 require "active_support/core_ext/hash/keys"
-require "fauxpaas/components"
+require "fauxpaas/components/paths"
+require "fauxpaas/components/backend_runner"
 require "fauxpaas/cap"
 
 module Fauxpaas
